requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
psycopg2
SQLAlchemy-Continuum